Make space() an ObsComponent method
