+++ gym_sts/spaces/observations/components/__init__.py
@@ -1,7 +1,7 @@
-from .campfire import CampfireStateObs  # noqa: F401
-from .card_reward import CardRewardStateObs  # noqa: F401
-from .combat import CombatStateObs  # noqa: F401
-from .combat_reward import CombatRewardState  # noqa: F401
-from .map import MapStateObs  # noqa: F401
+from .campfire import CampfireObs  # noqa: F401
+from .card_reward import CardRewardObs  # noqa: F401
+from .combat import CombatObs  # noqa: F401
+from .combat_reward import CombatRewardObs  # noqa: F401
+from .map import MapObs  # noqa: F401
 from .persistent import PersistentStateObs  # noqa: F401
-from .shop import ShopStateObs  # noqa: F401
+from .shop import ShopObs  # noqa: F401
