+++ gym_sts/spaces/observations/components/base.py
@@ -4,4 +4,9 @@ from abc import ABC, abstractmethod
 class ObsComponent(ABC):
     @abstractmethod
     def serialize(self):
-        raise RuntimeError("Not implemented")
+        raise NotImplementedError("Not implemented")
+
+    @staticmethod
+    @abstractmethod
+    def space():
+        raise NotImplementedError("Not implemented")
