+++ gym_sts/spaces/observations/components/card_reward.py
@@ -1,3 +1,4 @@
+from gym.spaces import Dict, Discrete, MultiBinary, Tuple
 from pydantic import parse_obj_as
 
 from gym_sts.spaces import constants
@@ -6,7 +7,7 @@ from gym_sts.spaces.observations import types
 from .base import ObsComponent
 
 
-class CardRewardStateObs(ObsComponent):
+class CardRewardObs(ObsComponent):
     def __init__(self, state: dict):
         # Sane defaults
         self.cards = []
@@ -24,6 +25,24 @@ class CardRewardStateObs(ObsComponent):
                 self.cards = parse_obj_as(list[types.Card], screen_state["cards"])
                 self.singing_bowl = screen_state["bowl_available"]
                 self.skippable = screen_state["skip_available"]
+
+    @staticmethod
+    def space():
+        return Dict(
+            {
+                # At most 4 cards may be offered (due to Question Card relic).
+                "cards": Tuple(
+                    (
+                        MultiBinary(constants.LOG_NUM_CARDS_WITH_UPGRADES),
+                        MultiBinary(constants.LOG_NUM_CARDS_WITH_UPGRADES),
+                        MultiBinary(constants.LOG_NUM_CARDS_WITH_UPGRADES),
+                        MultiBinary(constants.LOG_NUM_CARDS_WITH_UPGRADES),
+                    )
+                ),
+                "singing_bowl": Discrete(2),
+                "skippable": Discrete(2),
+            }
+        )
 
     def serialize(self) -> dict:
         serialized_cards = [
