+++ gym_sts/spaces/observations/components/combat_reward.py
@@ -1,10 +1,12 @@
+from gym.spaces import Dict, Discrete, MultiBinary, Tuple
+
 from gym_sts.spaces import constants
 from gym_sts.spaces.observations import types
 
 from .base import ObsComponent
 
 
-class CombatRewardState(ObsComponent):
+class CombatRewardObs(ObsComponent):
     def __init__(self, state: dict):
         # Sane defaults
         self.rewards: list[types.Reward] = []
@@ -27,6 +29,17 @@ class CombatRewardState(ObsComponent):
                 types.RelicReward(value=types.Relic(**relic))
                 for relic in screen_state["relics"]
             ]
+
+    @staticmethod
+    def space():
+        combat_reward_item = Dict(
+            {
+                "type": Discrete(constants.NUM_REWARD_TYPES),
+                # Could be a gold value, a relic ID, the color of a key, or a potion ID
+                "value": MultiBinary(constants.COMBAT_REWARD_LOG_MAX_ID),
+            }
+        )
+        return Tuple([combat_reward_item] * constants.MAX_NUM_REWARDS)
 
     @staticmethod
     def _parse_reward(reward: dict):
