+++ gym_sts/spaces/observations/spaces.py
@@ -1,0 +1,28 @@
+from gym.spaces import Dict, Discrete, MultiBinary, MultiDiscrete, Tuple
+
+from gym_sts.spaces import constants
+
+
+def generate_card_space():
+    # Generally beyond some number of cards you don't actually care
+    # how many cards you have
+    # But this could be optimized
+    return MultiDiscrete(
+        [constants.MAX_COPIES_OF_CARD] * constants.NUM_CARDS_WITH_UPGRADES
+    )
+
+
+def generate_effect_space():
+    return Tuple(
+        [Dict({"sign": Discrete(2), "value": MultiBinary([constants.LOG_MAX_EFFECT])})]
+        * constants.NUM_EFFECTS
+    )
+
+
+def generate_health_space():
+    return Dict(
+        {
+            "hp": MultiBinary(constants.LOG_MAX_HP),
+            "max_hp": MultiBinary(constants.LOG_MAX_HP),
+        }
+    )
