+++ gym_sts/spaces/observations/utils.py
@@ -1,8 +1,3 @@
-from gym_sts.spaces import constants
-
-from . import types
-
-
 def to_binary_array(n: int, digits: int) -> list[int]:
     array = [0] * digits
 
@@ -20,59 +15,3 @@ def to_binary_array(n: int, digits: int) -> list[int]:
         idx += 1
 
     return array
-
-
-def serialize_cards(cards: list[types.Card]) -> list[int]:
-    # TODO handle Searing Blow, which can be upgraded unlimited times
-    serialized = [0] * constants.NUM_CARDS_WITH_UPGRADES
-    for card in cards:
-        card_idx = card.serialize_discrete()
-
-        if serialized[card_idx] < constants.MAX_COPIES_OF_CARD:
-            serialized[card_idx] += 1
-
-    return serialized
-
-
-def serialize_health(hp: int, max_hp: int) -> dict[str, list[int]]:
-    return {
-        "hp": to_binary_array(hp, constants.LOG_MAX_HP),
-        "max_hp": to_binary_array(max_hp, constants.LOG_MAX_HP),
-    }
-
-
-def serialize_effects(effects: list) -> list[dict]:
-    serialized = []
-    effect_map = {effect["id"]: effect for effect in effects}
-
-    for effect in constants.ALL_EFFECTS:
-        encoding = {
-            "sign": 0,
-            "value": to_binary_array(0, constants.LOG_MAX_EFFECT),
-        }
-        if effect in effect_map:
-            value = effect_map[effect]["amount"]
-
-            if value < 0:
-                encoding["sign"] = 1
-                value = -value
-
-            encoding["value"] = to_binary_array(value, constants.LOG_MAX_EFFECT)
-
-        serialized.append(encoding)
-
-    return serialized
-
-
-def serialize_orbs(orbs: list) -> list:
-    serialized = [0] * constants.MAX_ORB_SLOTS
-    for i, orb in enumerate(orbs):
-        if "id" in orb:
-            orb_idx = constants.ALL_ORBS.index(orb["id"])
-        else:
-            # STS seems to have a bug where empty orbs sometimes have no ID
-            orb_idx = constants.ALL_ORBS.index("Empty")
-
-        serialized[i] = orb_idx
-
-    return serialized
